requirements.txt
absl-py==0.5.0
aiohttp==3.4.4
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
entrypoints==0.2.3
f==0.0.1
fake-useragent==0.1.11
Flask==1.0.2
gast==0.2.0
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.8.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
multidict==4.5.2
mysqlclient==1.3.13
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.14.5
pandas==0.23.4
pandocfilters==1.4.2
parse==1.9.0
parsel==1.5.0
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
pipenv==2018.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
pyee==5.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
pymongo==3.7.2
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.1
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
queuelib==1.5.0
requests==2.19.1
requests-html==0.9.0
scikit-learn==0.20.2
scipy==1.1.0
selenium==3.14.1
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
static3==0.7.0
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
yarl==1.2.6
zope.interface==4.5.0
